Move test case parsing into separate class
